Try and remove rdfStore (or update performance)
